Move ACMS chapter normalizer to gateway helper
Chapter normalization of ACMS's raw legacy codes (7A/7N/09/AC) is
a gateway-adapter concern, not a CAMS domain concept —
reconsidered after the prior move to common/src/cams/cases.ts.

- Move normalizeAcmsCaseChapter to a new
  backend/lib/adapters/gateways/acms/acms.gateway.helper.ts,
  matching the existing dxtr.gateway.helper.ts pattern for
  source-system-specific parsing/normalization logic.
- CaseChapter/VALID_CASE_CHAPTERS remain in
  common/src/cams/cases.ts, since that vocabulary is genuinely
  shared across the ACMS normalizer, the DXTR validator, the
  repository guard, and the search validator.
- Move the normalizer's unit tests to a new
  acms.gateway.helper.test.ts alongside it.
