+++ common/src/cams/cases.test.ts
@@ -9,7 +9,6 @@ import {
   isLeadCase,
   isTransferredCase,
   MatchType,
-  normalizeAcmsCaseChapter,
   ScoreBreakdown,
   SearchMetadata,
   SyncedCase,
@@ -227,35 +226,6 @@ describe('cases common functions tests', () => {
 
     test('should throw for malformed case ID with no hyphens', () => {
       expect(() => getCaseNumber('invalid')).toThrow('Invalid case ID: invalid');
-    });
-  });
-
-  describe('normalizeAcmsCaseChapter', () => {
-    test.each(['7A', '7N'])("normalizes '%s' to '7'", (rawChapter) => {
-      expect(normalizeAcmsCaseChapter(rawChapter)).toBe('7');
-    });
-
-    test("normalizes '09' to '9'", () => {
-      expect(normalizeAcmsCaseChapter('09')).toBe('9');
-    });
-
-    test.each(['7', '9', '11', '12', '13', '15'])(
-      "passes valid chapter '%s' through unchanged",
-      (chapter) => {
-        expect(normalizeAcmsCaseChapter(chapter)).toBe(chapter);
-      },
-    );
-
-    test('trims whitespace', () => {
-      expect(normalizeAcmsCaseChapter(' 7 ')).toBe('7');
-    });
-
-    test("throws for unrecognized chapter 'AC'", () => {
-      expect(() => normalizeAcmsCaseChapter('AC')).toThrow('Invalid ACMS chapter value: AC');
-    });
-
-    test('throws for an empty string', () => {
-      expect(() => normalizeAcmsCaseChapter('')).toThrow('Invalid ACMS chapter value:');
     });
   });
 });
